:dizzy: よく使う関数をGlobalFunctions.swiftに移した #58
